# swagger/Trainingsplanner-postgres.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "laplanner",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/appointments/{appointmentId}/modules/{moduleId}": {
      "delete": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_DeleteAppointmentWithModule",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "202": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/trainingsexercises/modules/{trainingsModuleId}": {
      "get": {
        "tags": [
          "TrainingsExercise"
        ],
        "summary": "Returns the TrainingsExercise with the given Id",
        "operationId": "TrainingsExercise_ReadExercisesByModuleId",
        "parameters": [
          {
            "name": "trainingsModuleId",
            "in": "path",
            "required": true,
            "description": "Id of the required TrainingsModule",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The TrainingsExercise",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsExerciseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/modules/{moduleId}/tags/{tagId}": {
      "delete": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Deletes the TrainingsModuleTag with the given Id from the given Module",
        "operationId": "TrainingsModule_DeleteTagByModuleId",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "description": "The TrainingsModuleId (from HTTP-Body)",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "description": "The TagId(from HTTP-Body)",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/modules/{id}": {
      "get": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Returns all TrainingsModules",
        "operationId": "TrainingsModule_GetTrainingsModuleById",
        "parameters": [
          {
            "name": "trainingscModuleId",
            "in": "query",
            "description": "Id of the needed TrainingsModule",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "The TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tags": {
      "get": {
        "tags": [
          "TrainingsModuleTag"
        ],
        "operationId": "TrainingsModuleTag_GetAllTags",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsModuleTagDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TrainingsModuleTag"
        ],
        "operationId": "TrainingsModuleTag_CreateTag",
        "requestBody": {
          "x-name": "trainingsModuleTagDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsModuleTagDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleTagDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TrainingsModuleTag"
        ],
        "operationId": "TrainingsModuleTag_UpdateTag",
        "requestBody": {
          "x-name": "trainingsModuleTagDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsModuleTagDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleTagDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TrainingsModuleTag"
        ],
        "operationId": "TrainingsModuleTag_DeleteTag",
        "requestBody": {
          "x-name": "trainingsModuleTagDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsModuleTagDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/{id}": {
      "get": {
        "tags": [
          "TrainingsModuleTag"
        ],
        "operationId": "TrainingsModuleTag_GetTagById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleTagDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TrainingsGroup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "trainingsGroupsApplicationUsers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsGroupApplicationUser"
            }
          },
          "trainingsAppointments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointment"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsGroupApplicationUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "applicationUserId": {
            "type": "string",
            "nullable": true
          },
          "isTrainer": {
            "type": "boolean"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsAppointment": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsAppointmentTrainingsModule": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsAppointmentId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsAppointment": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsAppointment"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsModule": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "$ref": "#/components/schemas/TrainingsDifficulty"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsDifficulty": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "UNKNOWN",
          "EASY",
          "MEDIUM",
          "HARD",
          "IMPOSSIBLE"
        ],
        "enum": [
          0,
          10,
          20,
          30,
          40
        ]
      },
      "TrainingsModuleTrainingsExercise": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsExerciesId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsExercise": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsExercise"
              }
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsExercise": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "repetitions": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsModuleTrainingsModuleTag": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsModuleTagId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModuleTag": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModuleTag"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsModuleTag": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TrainingsModuleDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "$ref": "#/components/schemas/TrainingsDifficulty"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "TrainingsExerciseDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "repetitions": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          }
        }
      },
...
      "TrainingsModuleTagDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          }
        }
      },
...
    }
  }
}